Create DiscordCallAdmin.cs
# Tested on oxide.ext.discord 3:
- Oxide: v2.0.6333
- Discord extension: v3.1.2

If used with the aforementioned version, invoking the calladmin command will result in an error:
```
 "message": "Value \"0\" is not a valid enum value.
```

Discord's API now requires that the `default_auto_archive_duration` field be set to a valid enum value when creating channels. Since the value 0 is no longer accepted, I’ve updated the script to specify a valid duration, ensuring that channel creation requests are successful.